// src/api/nvd/cves/2.0/schema/response.ts
export type CveId = string;
// ...
export interface DefCveItem {
  cve: CveItem;
}
export interface CveItem {
  id: CveId;
  sourceIdentifier?: string;
  vulnStatus?: string;
  published: string;
  lastModified: string;
  evaluatorComment?: string;
  evaluatorSolution?: string;
  evaluatorImpact?: string;
  cisaExploitAdd?: string;
  cisaActionDue?: string;
  cisaRequiredAction?: string;
  cisaVulnerabilityName?: string;
  cveTags?: {
    /**
     * The email address or UUID of the source that contributed the information
     */
    sourceIdentifier?: string;
    tags?: ("unsupported-when-assigned" | "exclusively-hosted-service" | "disputed")[];
    [k: string]: unknown;
  }[];
  /**
   * @minItems 1
   */
  descriptions: [LangString, ...LangString[]];
  references: Reference[];
  /**
   * Metric scores for a vulnerability as found on NVD.
   */
  metrics?: {
    /**
     * CVSS V4.0 score.
     */
    cvssMetricV40?: CvssV40[];
    /**
     * CVSS V3.1 score.
     */
    cvssMetricV31?: CvssV31[];
    /**
     * CVSS V3.0 score.
     */
    cvssMetricV30?: CvssV30[];
    /**
     * CVSS V2.0 score.
     */
    cvssMetricV2?: CvssV2[];
    [k: string]: unknown;
  };
  weaknesses?: Weakness[];
  configurations?: Config[];
  vendorComments?: VendorComment[];
  [k: string]: unknown;
}
export interface LangString {
  lang: string;
  value: string;
}
export interface Reference {
  url: string;
  source?: string;
  tags?: string[];
  [k: string]: unknown;
}
export interface CvssV40 {
  source: string;
  type: "Primary" | "Secondary";
  cvssData: JSONSchemaForCommonVulnerabilityScoringSystemVersion40;
}
export interface CvssV31 {
  source: string;
  type: "Primary" | "Secondary";
  cvssData: JSONSchemaForCommonVulnerabilityScoringSystemVersion31;
  exploitabilityScore?: DefSubscore;
  impactScore?: DefSubscore;
}
// ...
export interface CvssV30 {
  source: string;
  type: "Primary" | "Secondary";
  cvssData: JSONSchemaForCommonVulnerabilityScoringSystemVersion30;
  exploitabilityScore?: DefSubscore;
  impactScore?: DefSubscore;
}
// ...
export interface CvssV2 {
  source: string;
  type: "Primary" | "Secondary";
  cvssData: JSONSchemaForCommonVulnerabilityScoringSystemVersion20;
  baseSeverity?: string;
  exploitabilityScore?: DefSubscore;
  impactScore?: DefSubscore;
  acInsufInfo?: boolean;
  obtainAllPrivilege?: boolean;
  obtainUserPrivilege?: boolean;
  obtainOtherPrivilege?: boolean;
  userInteractionRequired?: boolean;
}
// ...
export interface Weakness {
  source: string;
  type: string;
  /**
   * @minItems 0
   */
  description: LangString[];
}
export interface Config {
  operator?: "AND" | "OR";
  negate?: boolean;
  nodes: Node[];
  [k: string]: unknown;
}
// ...
export interface VendorComment {
  organization: string;
  comment: string;
  lastModified: string;
}